+++ src/c_yui_gobj_tree_js.js
@@ -59,6 +59,7 @@ import {
 } from '@antv/g6';
 
 import { ensure_drag_canvas_patch } from "./g6_drag_canvas_touch.js";
+import { ensure_pinch_zoom_patch } from "./g6_touch_gestures.js";
 import { yui_is_dark, yui_theme_now, yui_watch_theme } from "./yui_theme.js";
 
 /***************************************************************
@@ -305,6 +306,7 @@ function mt_create(gobj)
     build_ui(gobj);
 
     ensure_drag_canvas_patch();
+    ensure_pinch_zoom_patch();
 
     if(!__layout_registered__) {
         register(ExtensionCategory.LAYOUT, 'gobj-tree-v', GobjTreeVLayout);